fix (#33) : 네이버 체크스타일 반영
