+++ src/main/java/com/limito/order/cart/domain/dto/resellproduct/AddCartResellRequestV1.java
@@ -1,4 +1,4 @@
-package com.limito.order.cart.domain.dto.resellProduct;
+package com.limito.order.cart.domain.dto.resellproduct;
 
 import java.util.UUID;
