+++ src/main/java/com/limito/order/order/domain/dto/feignclient/resell/dto/request/StockReduceRequest.java
@@ -1,4 +1,4 @@
-package com.limito.order.order.domain.dto.feignClient.resell.dto.request;
+package com.limito.order.order.domain.dto.feignclient.resell.dto.request;
 
 import java.util.UUID;
