+++ src/main/java/com/limito/order/order/service/OrderInternalServiceV1.java
@@ -13,12 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 import com.limito.common.exception.AppException;
 import com.limito.order.cart.service.CartServiceV1;
 import com.limito.order.common.OrderStatus;
-import com.limito.order.common.feignClient.LimitedFeignClient;
-import com.limito.order.common.feignClient.ResellFeignClient;
-import com.limito.order.order.domain.dto.feignClient.limited.ReduceStockProductRequestV1;
-import com.limito.order.order.domain.dto.feignClient.limited.ReduceStockRequestV1;
-import com.limito.order.order.domain.dto.feignClient.resell.dto.request.StockReduceRequest;
-import com.limito.order.order.domain.dto.feignClient.resell.dto.response.InternalResponse;
+import com.limito.order.common.feignclient.LimitedFeignClient;
+import com.limito.order.common.feignclient.ResellFeignClient;
+import com.limito.order.order.domain.dto.feignclient.limited.ReduceStockProductRequestV1;
+import com.limito.order.order.domain.dto.feignclient.limited.ReduceStockRequestV1;
+import com.limito.order.order.domain.dto.feignclient.resell.dto.request.StockReduceRequest;
+import com.limito.order.order.domain.dto.feignclient.resell.dto.response.InternalResponse;
 import com.limito.order.order.domain.mapper.OrderMapper;
 import com.limito.order.order.domain.model.Order;
 import com.limito.order.order.domain.model.OrderItem;
